Messed with code in such a way as to totally break it. Need to rethink how pages work
Each page should essentially know where it starts, and how many objects (that's easy to count in the children array). Scratch that -- mid commit thought change. Just one global count, and the number of elements on the current page. Or better yet, the carousel is just 3 slides infinitely looping, but each slide is replaced before it's viewed.